Rename test files. Start working on Proxy DP
