+++ README.md
@@ -1,4 +1,8 @@
 # Studying design patterns using TypeScript and Jest
 
+All v1.0 modules and specs are here for historical reasons only.
+All v2.0 modules and specs supersede v1.0 ones.
+Modules and specs without versions should be considered legacy v1.0 type.
+
 ### Changelog:
 - 10dec2023: adapter-v2.0: `src/design-patterns/Adapter`
