+++ src/design-patterns/Adapter/adapter-v1.0.spec.ts
