+++ src/design-patterns/Adapter/adapter-v2.0.spec.ts
