+++ src/design-patterns/Proxy/proxy-v2.0.ts
@@ -1,0 +1,5 @@
+export class Subject {
+  request(): void {
+    console.log("RealSubject: Handling request.")
+  }
+}
